Centered the fractals, colorized them
I have coded some adjustments for the fractals in order for them to be in the middle of the canvas element (their length now also varies, depending on the level). Also I have removed the option to choose size, because it can be used as a freeway for the curve to escape - it is guaranteed that the image will look good - that is of course that you pick the right color combination, as I have added font and background full support. Also I have removed a paragraph which contained current screen size, as it was now unncessary. Also I have corrected the recursive approach to linear (in the construction of the L-system), in order to minimize function calls