Download option, bug fixing, additional UX
I have added a download option for all the fractal images. Also the background has been correctly adjusted to be apparent in the image. I have added some performance checks for the fractal (if a new one is same as old, it won't be regenerated). Some UX for the buttons, like hovering, background enable/disable option